Refactoring dao & module (all refreshing, needs improvement)
