+++ app/src/main/java/ru/netology/nmedia/dao/PostDao.kt
@@ -11,10 +11,13 @@ import ru.netology.nmedia.entity.PostEntity
 
 @Dao
 interface PostDao {
+//    @Query("SELECT * FROM PostEntity WHERE viewed = 0 ORDER BY id DESC")
+//    fun getAll(): Flow<List<PostEntity>>
+
     @Query("SELECT * FROM PostEntity WHERE viewed = 0 ORDER BY id DESC")
     fun getAll(): Flow<List<PostEntity>>
 
-    @Query("SELECT * FROM PostEntity ORDER BY id DESC")
+    @Query("SELECT * FROM PostEntity WHERE viewed = 0 ORDER BY id DESC")
     fun getPagingSource(): PagingSource<Int, PostEntity>
 
     @Insert(onConflict = REPLACE)
