+++ app/src/main/java/ru/netology/nmedia/dao/PostRemoteKeyDao.kt
@@ -9,10 +9,10 @@ import ru.netology.nmedia.entity.PostRemoteKeyEntity
 @Dao
 interface PostRemoteKeyDao {
 
-    @Query("SELECT max(`key`) FROM PostRemoteKeyEntity")
+    @Query("SELECT MAX(`key`) FROM PostRemoteKeyEntity")
     suspend fun max(): Long?
 
-    @Query("SELECT min(`key`) FROM PostRemoteKeyEntity")
+    @Query("SELECT MIN(`key`) FROM PostRemoteKeyEntity")
     suspend fun min(): Long?
 
     @Insert(onConflict = REPLACE)
