--- app/src/main/java/ru/netology/nmedia/module/DaoModule.kt
@@ -1,19 +1,0 @@
-package ru.netology.nmedia.module
-
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import ru.netology.nmedia.dao.PostDao
-import ru.netology.nmedia.dao.PostRemoteKeyDao
-import ru.netology.nmedia.db.AppDb
-
-@InstallIn(SingletonComponent::class)
-@Module
-object DaoModule {
-    @Provides
-    fun providePostDao(db: AppDb): PostDao = db.postDao()
-
-    @Provides
-    fun providePostRemoteKeyDao(db: AppDb): PostRemoteKeyDao = db.postRemoteKeyDao()
-}
