+++ app/src/main/java/ru/netology/nmedia/module/DbModule.kt
@@ -1,4 +1,4 @@
-package ru.netology.nmedia.db
+package ru.netology.nmedia.module
 
 import android.content.Context
 import androidx.room.Room
@@ -8,6 +8,8 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import ru.netology.nmedia.dao.PostDao
+import ru.netology.nmedia.dao.PostRemoteKeyDao
+import ru.netology.nmedia.db.AppDb
 import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
@@ -27,4 +29,8 @@ class DbModule {
     fun providePostDao(
         appDb: AppDb,
     ): PostDao = appDb.postDao()
+
+    @Provides
+    fun providesPostRemoteKeyDao(appDb: AppDb):
+            PostRemoteKeyDao = appDb.postRemoteKeyDao()
 }
